Extract clip tests
